feat: Added features for deriving
